refactor use asset
